Consolidate exceptions
