+++ pygerduty/__init__.py
@@ -8,7 +8,7 @@ import re
 import six
 from six import string_types
 from six.moves import urllib
-
+from exceptions import Error, IntegrationAPIError
 
 __author__ = "Mike Cugini <cugini@dropbox.com>"
 from .version import __version__, version_info  # noqa
@@ -22,39 +22,6 @@ ISO8601_FORMAT = "%Y-%m-%dT%H:%M:%SZ"
 # TODO:
 # Support for Log Entries
 # Support for Reports
-
-
-class Error(Exception):
-    pass
-
-
-class IntegrationAPIError(Error):
-    def __init__(self, message, event_type):
-        self.event_type = event_type
-        self.message = message
-
-    def __str__(self):
-        return "Creating {0} event failed: {1}".format(self.event_type,
-                                                       self.message)
-
-
-class BadRequest(Error):
-    def __init__(self, payload, *args, **kwargs):
-        # Error Reponses don't always contain all fields.
-        # Sane defaults must be set.
-        self.code = payload.get("error", {}).get('code', 99999)
-        self.errors = payload.get("error", {}).get('errors', [])
-        self.message = payload.get("error", {}).get('message', str(payload))
-
-        Error.__init__(self, *args, **kwargs)
-
-    def __str__(self):
-        return "{0} ({1}): {2}".format(
-            self.message, self.code, self.errors)
-
-
-class NotFound(Error):
-    pass
 
 
 class Collection(object):
